refactor(desktop): one script runner into the preview guest page, not a tour-only one
