+++ apps/desktop/src/app/chat/right-rail/preview-tour.ts
@@ -21,7 +21,7 @@ import driverIife from 'driver.js/dist/driver.js.iife.js?raw'
 import { collectTourTargets } from '@/lib/tour/collect-targets'
 import { runTourEngine, type TourAction, type TourResult } from '@/lib/tour/engine'
 
-import { activeTourRunner } from './preview-tour-runner'
+import { activePreviewScriptRunner } from './preview-script-runner'
 
 /** Build the idempotent inject-and-run script for one tour action. */
 function buildTourScript(action: TourAction): string {
@@ -51,7 +51,7 @@ function buildTourScript(action: TourAction): string {
 
 /** Run one tour action in the ACTIVE preview tab's page. */
 export async function runPreviewTour(action: TourAction): Promise<TourResult> {
-  const run = activeTourRunner()
+  const run = activePreviewScriptRunner()
 
   if (!run) {
     return { error: 'No live page is open in the preview pane — open one first.', success: false }
